remove unused non JIT code paths
