+++ front/lib/api/assistant/actions/conversation/include_file.ts
@@ -107,7 +107,6 @@ export class ConversationIncludeFileAction extends BaseAction {
           forceFullCSVInclude: true, // We are using JIT so if we are rendering a CSV file we want to include it in full
           model,
           title: f.title,
-          textBytes: null,
           contentFragmentVersion: f.contentFragmentVersion,
         });
