+++ front/lib/api/assistant/generation.ts
@@ -18,7 +18,6 @@ import {
   isContentFragmentMessageTypeModel,
   isContentFragmentType,
   isRetrievalConfiguration,
-  isTextContent,
   isUserMessageType,
   isWebsearchConfiguration,
   Ok,
@@ -29,17 +28,13 @@ import moment from "moment-timezone";
 import { citationMetaPrompt } from "@app/lib/api/assistant/citations";
 import { getAgentConfigurations } from "@app/lib/api/assistant/configuration";
 import {
-  isJITActionsEnabled,
-  renderConversationForModelJIT,
-} from "@app/lib/api/assistant/jit_actions";
-import {
   getTextContentFromMessage,
   getTextRepresentationFromMessages,
 } from "@app/lib/api/assistant/utils";
 import { getVisualizationPrompt } from "@app/lib/api/assistant/visualization";
 import type { Authenticator } from "@app/lib/auth";
-import { renderContentFragmentForModel } from "@app/lib/resources/content_fragment_resource";
-import { tokenCountForTexts, tokenSplit } from "@app/lib/tokenization";
+import { renderLightContentFragmentForModel } from "@app/lib/resources/content_fragment_resource";
+import { tokenCountForTexts } from "@app/lib/tokenization";
 import logger from "@app/logger/logger";
 
 /**
@@ -168,54 +163,7 @@ export async function constructPromptMultiActions(
  * Model conversation rendering
  */
 
-export async function renderConversationForModel(
-  auth: Authenticator,
-  {
-    conversation,
-    model,
-    prompt,
-    allowedTokenCount,
-    excludeActions,
-    excludeImages,
-  }: {
-    conversation: ConversationType;
-    model: ModelConfigurationType;
-    prompt: string;
-    allowedTokenCount: number;
-    excludeActions?: boolean;
-    excludeImages?: boolean;
-  }
-): Promise<
-  Result<
-    {
-      modelConversation: ModelConversationTypeMultiActions;
-      tokensUsed: number;
-    },
-    Error
-  >
-> {
-  if (!isJITActionsEnabled()) {
-    return renderConversationForModelMultiActions({
-      conversation,
-      model,
-      prompt,
-      allowedTokenCount,
-      excludeActions,
-      excludeImages,
-    });
-  } else {
-    return renderConversationForModelJIT({
-      conversation,
-      model,
-      prompt,
-      allowedTokenCount,
-      excludeActions,
-      excludeImages,
-    });
-  }
-}
-
-async function renderConversationForModelMultiActions({
+export async function renderConversationForModel({
   conversation,
   model,
   prompt,
@@ -241,36 +189,40 @@ async function renderConversationForModelMultiActions({
   const now = Date.now();
   const messages: ModelMessageTypeMultiActions[] = [];
 
-  // Render loop.
-  // Render all messages and all actions.
+  // Render loop: dender all messages and all actions.
   for (const versions of conversation.content) {
     const m = versions[versions.length - 1];
 
     if (isAgentMessageType(m)) {
       const actions = removeNulls(m.actions);
 
-      // This array is 2D, because we can have multiple calls per agent message (parallel calls).
-
-      const steps = [] as Array<{
-        contents: string[];
-        actions: Array<{
-          call: FunctionCallType;
-          result: FunctionMessageTypeModel;
-        }>;
-      }>;
+      // This is a record of arrays, because we can have multiple calls per agent message (parallel
+      // calls).  Actions all have a step index which indicates how they should be grouped but some
+      // actions injected by `getEmulatedAgentMessageActions` have a step index of `-1`. We
+      // therefore group by index, then order and transform in a 2D array to present to the model.
+      const stepByStepIndex = {} as Record<
+        string,
+        {
+          contents: string[];
+          actions: Array<{
+            call: FunctionCallType;
+            result: FunctionMessageTypeModel;
+          }>;
+        }
+      >;
 
       const emptyStep = () =>
         ({
           contents: [],
           actions: [],
-        }) satisfies (typeof steps)[number];
+        }) satisfies (typeof stepByStepIndex)[number];
 
       for (const action of actions) {
         const stepIndex = action.step;
-        steps[stepIndex] = steps[stepIndex] || emptyStep();
+        stepByStepIndex[stepIndex] = stepByStepIndex[stepIndex] || emptyStep();
         // All these calls (except `conversation_include_files_action` are not async so we're not
         // doing a Promise.all for now but might need to be reconsiderd in the future.
-        steps[stepIndex].actions.push({
+        stepByStepIndex[stepIndex].actions.push({
           call: action.renderForFunctionCall(),
           result: await action.renderForMultiActionsModel({
             conversation,
@@ -280,11 +232,16 @@ async function renderConversationForModelMultiActions({
       }
 
       for (const content of m.rawContents) {
-        steps[content.step] = steps[content.step] || emptyStep();
+        stepByStepIndex[content.step] =
+          stepByStepIndex[content.step] || emptyStep();
         if (content.content.trim()) {
-          steps[content.step].contents.push(content.content);
+          stepByStepIndex[content.step].contents.push(content.content);
         }
       }
+
+      const steps = Object.entries(stepByStepIndex)
+        .sort(([a], [b]) => Number(a) - Number(b))
+        .map(([, step]) => step);
 
       if (excludeActions) {
         // In Exclude Actions mode, we only render the last step that has content.
@@ -372,14 +329,11 @@ async function renderConversationForModelMultiActions({
         ],
       });
     } else if (isContentFragmentType(m)) {
-      const res = await renderContentFragmentForModel(m, conversation, model, {
-        excludeImages: Boolean(excludeImages),
-      });
-
-      if (res.isErr()) {
-        return new Err(res.error);
-      }
-      messages.push(res.value);
+      messages.push(
+        await renderLightContentFragmentForModel(m, conversation, model, {
+          excludeImages: Boolean(excludeImages),
+        })
+      );
     } else {
       assertNever(m);
     }
@@ -390,7 +344,6 @@ async function renderConversationForModelMultiActions({
     [prompt, ...getTextRepresentationFromMessages(messages)],
     model
   );
-
   if (res.isErr()) {
     return new Err(res.error);
   }
@@ -398,14 +351,12 @@ async function renderConversationForModelMultiActions({
   const [promptCount, ...messagesCount] = res.value;
 
   // We initialize `tokensUsed` to the prompt tokens + a bit of buffer for message rendering
-  // approximations, 64 tokens seems small enough and ample enough.
+  // approximations.
   const tokensMargin = 1024;
   let tokensUsed = promptCount + tokensMargin;
 
   // Go backward and accumulate as much as we can within allowedTokenCount.
   const selected: ModelMessageTypeMultiActions[] = [];
-  const truncationMessage = `... (content truncated)`;
-  const approxTruncMsgTokenCount = truncationMessage.length / 3;
 
   // Selection loop.
   for (let i = messages.length - 1; i >= 0; i--) {
@@ -416,57 +367,12 @@ async function renderConversationForModelMultiActions({
     if (tokensUsed + c <= allowedTokenCount) {
       tokensUsed += c;
       selected.unshift(currentMessage);
-    } else if (
-      // When a content fragment has more than the remaining number of tokens, we split it.
-      isContentFragmentMessageTypeModel(currentMessage) &&
-      // Allow at least tokensMargin tokens in addition to the truncation message.
-      tokensUsed + approxTruncMsgTokenCount + tokensMargin < allowedTokenCount
-    ) {
-      const remainingTokens =
-        allowedTokenCount - tokensUsed - approxTruncMsgTokenCount;
-
-      const updatedContent = [];
-      for (const c of currentMessage.content) {
-        if (!isTextContent(c)) {
-          // If there is not enough room and it's an image, we simply ignore it.
-          continue;
-        }
-
-        // Remove only if it ends with "</attachment>".
-        const textWithoutClosingAttachmentTag = c.text.replace(
-          /<\/attachment>$/,
-          ""
-        );
-
-        const contentRes = await tokenSplit(
-          textWithoutClosingAttachmentTag,
-          model,
-          remainingTokens
-        );
-        if (contentRes.isErr()) {
-          return new Err(contentRes.error);
-        }
-
-        updatedContent.push({
-          ...c,
-          text: `${contentRes.value}${truncationMessage}</attachment>`,
-        });
-      }
-
-      selected.unshift({
-        ...currentMessage,
-        content: updatedContent,
-      });
-
-      tokensUsed += remainingTokens;
-      break;
     } else {
       break;
     }
   }
 
-  // Merging loop.
-  // Merging content fragments into the upcoming user message.
+  // Merging loop: merging content fragments into the upcoming user message.
   // Eg: [CF1, CF2, UserMessage, AgentMessage] => [CF1-CF2-UserMessage, AgentMessage]
   for (let i = selected.length - 1; i >= 0; i--) {
     const cfMessage = selected[i];
